+++ src/bin/wasm-tools/print.rs
@@ -1,0 +1,29 @@
+use anyhow::{Context, Result};
+use std::path::PathBuf;
+use structopt::StructOpt;
+
+/// Print the textual form of a WebAssembly binary.
+#[derive(StructOpt)]
+pub struct Opts {
+    /// Input WebAssembly file to print.
+    input: PathBuf,
+
+    /// An optional output file to place the output into.
+    ///
+    /// If not specified then the wasm file is printed to standard output.
+    #[structopt(short = "o", long)]
+    output: Option<PathBuf>,
+}
+
+impl Opts {
+    pub fn run(&self) -> Result<()> {
+        let wat = wasmprinter::print_file(&self.input)?;
+        if let Some(output) = &self.output {
+            std::fs::write(&output, wat).context(format!("failed to write {:?}", output))?;
+        } else {
+            println!("{}", wat);
+        }
+
+        Ok(())
+    }
+}
